# docs/swagger.yaml
definitions:
...
  api.CreateMaterialRequest:
    properties:
      course_id:
        type: integer
      description:
        type: string
      title:
        type: string
    required:
    - course_id
    - description
    - title
    type: object
...
  api.UpdateMaterialRequest:
    properties:
      course_id:
        type: integer
      description:
        type: string
      material_id:
        minimum: 1
        type: integer
      title:
        type: string
    required:
    - course_id
    - description
    - material_id
    - title
    type: object
...
  api.createRequestRequest:
    properties:
      course_id:
        type: integer
      is_accepted:
        type: boolean
      is_active:
        type: boolean
      is_declined:
        type: boolean
      is_pending:
        type: boolean
      user_id:
        type: integer
    type: object
  api.createResourceRequest:
    properties:
      content_url:
        type: string
      material_id:
        type: integer
      title:
        type: string
      type:
        type: string
    type: object
...
  api.updateRequest:
    type: object
info:
  contact: {}
paths:
...
  /material:
    post:
      consumes:
      - application/json
      description: Create a new material
      operationId: create-material
      parameters:
      - description: Create Material Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.CreateMaterialRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Create a new material
  /material/{course_id}:
    get:
      consumes:
      - application/json
      description: Get materials for a course
      operationId: get-materials
      parameters:
      - description: Course ID
        in: path
        name: course_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get materials for a course
  /material/{material_id}:
    delete:
      consumes:
      - application/json
      description: Delete a material
      operationId: delete-material
      parameters:
      - description: Material ID
        in: path
        name: material_id
        required: true
        type: integer
      - description: Course ID
        in: path
        name: course_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Delete a material
    put:
      consumes:
      - application/json
      description: Update a material
      operationId: update-material
      parameters:
      - description: Material ID
        in: path
        name: material_id
        required: true
        type: integer
      - description: Update Material Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.UpdateMaterialRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Update a material
  /request/{request_id}:
    get:
      description: Get a request by its ID
      parameters:
      - description: Request ID
        in: path
        name: request_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get a request by ID
  /request/{user_id}/{request_id}:
    delete:
      description: Delete a request by student and request ID
      parameters:
      - description: Student ID
        in: path
        name: student_id
        required: true
        type: integer
      - description: Request ID
        in: path
        name: request_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Delete a request
  /request/update:
    put:
      consumes:
      - application/json
      description: Update a request with the provided parameters
      parameters:
      - description: Requested data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.updateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Update a request
  /requests:
    get:
      description: List requests based on provided parameters
      parameters:
      - description: Student ID
        in: query
        name: student_id
        type: integer
      - description: Teacher ID
        in: query
        name: teacher_id
        type: integer
      - description: Course ID
        in: query
        name: course_id
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: List requests
    post:
      consumes:
      - application/json
      description: Create a new request with the given parameters
      parameters:
      - description: Request data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.createRequestRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Create a new request
  /resource:
    post:
      consumes:
      - application/json
      description: Creates a new resource
      operationId: create-resource
      parameters:
      - description: Create Resource Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.createResourceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Create a new resource
  /resource/{resource_id}:
    delete:
      description: Deletes a resource
      operationId: delete-resource
      parameters:
      - description: Resource ID
        in: path
        name: resource_id
        required: true
        type: integer
      - description: Material ID
        in: path
        name: material_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Delete a resource
  /resource/{resource_id}/{material_id}:
    get:
      description: Retrieves a resource
      operationId: get-resource
      parameters:
      - description: Material ID
        in: path
        name: material_id
        required: true
        type: integer
      - description: Resource ID
        in: path
        name: resource_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get a resource
...
swagger: "2.0"